Create list type that contains list and size
This fixes a bug in the prior implementation where one of the recursive
cases in subsequencesOfSize was not decrementing the list size.

I never did find a test case that could expose the error, which bothers
me a bit, but storing the size in the type is less error prone.